requirements.txt
black==25.9.0
blinker==1.9.0
click==8.3.0
Flask==3.1.2
flask-cors==5.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytokens==0.2.0
Werkzeug==3.1.3
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==7.0.0
requests==2.32.3
